backend/requirements.txt
django
djangorestframework
gunicorn
asgiref
python-decouple
pytz
requests
django-jazzmin
drf_yasg
psycopg2-binary
bs4
Pillow